# api/docs/swagger.yaml
definitions:
  models.Author:
    properties:
      address:
        type: string
      age:
        type: integer
      birth_date:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  models.AuthorsResponse:
    properties:
      authors:
        items:
          $ref: '#/definitions/models.Author'
        type: array
      count:
        type: integer
    type: object
  models.Clinic:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.ClinicsResponse:
    properties:
      clinics:
        items:
          $ref: '#/definitions/models.Clinic'
        type: array
      count:
        type: integer
    type: object
  models.CreateAuthor:
    properties:
      address:
        type: string
      birth_date:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  models.CreateClinic:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  models.UpdateAuthor:
    properties:
      address:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
  models.UpdateClinic:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: Online doctor appointments and drug orders
  title: ShifoLink
  version: 1.0.0
paths:
  /author:
    post:
      consumes:
      - application/json
      description: Create a new author
      parameters:
      - description: author data
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/models.CreateAuthor'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Author'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a new author
      tags:
      - author
  /author/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Author
      parameters:
      - description: author id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete Author
      tags:
      - author
    get:
      consumes:
      - application/json
      description: Get auhtor by id
      parameters:
      - description: author
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Author'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get author by id
      tags:
      - author
    put:
      consumes:
      - application/json
      description: Update author by id
      parameters:
      - description: author id
        in: path
        name: id
        required: true
        type: string
      - description: author
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAuthor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Author'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update author by id
      tags:
      - author
  /authors:
    get:
      consumes:
      - application/json
      description: Get authors list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthorsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get authors list
      tags:
      - author
  /clinic:
    post:
      consumes:
      - application/json
      description: Create a new clinic
      parameters:
      - description: clinic data
        in: body
        name: clinic
        required: true
        schema:
          $ref: '#/definitions/models.CreateClinic'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Clinic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a new clinic
      tags:
      - clinic
  /clinic/{id}:
    delete:
      consumes:
      - application/json
      description: Delete clinic
      parameters:
      - description: clinic id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete clinic
      tags:
      - clinic
    get:
      consumes:
      - application/json
      description: Get clinic by id
      parameters:
      - description: clinic
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Clinic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get clinic by id
      tags:
      - clinic
    put:
      consumes:
      - application/json
      description: Update clinic by id
      parameters:
      - description: clinic id
        in: path
        name: id
        required: true
        type: string
      - description: clinic
        in: body
        name: clinic
        required: true
        schema:
          $ref: '#/definitions/models.UpdateClinic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Clinic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update clinic by id
      tags:
      - clinic
...
  /clinics:
    get:
      consumes:
      - application/json
      description: Get clinics list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ClinicsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get clinics list
      tags:
      - clinic
...
swagger: "2.0"